+++ Movies.h
@@ -6,13 +6,13 @@ using namespace std;
 class Movies:public Media{
 public:
   void setDirector(char*);
-  void setDuration(unsigned long);
+  void setDuration(int);
   void setRating(int);
   char* getDirector();
-  unsigned long getDuration();
+  int getDuration();
   int getRating();
 
   char director[80];
-  unsigned long duration;
+  int duration;
   int rating;
 };
