+++ Music.cpp
@@ -8,7 +8,7 @@ void Music::setArtist(char* artistTemp) {
   strncpy(artist, artistTemp, 79);
 }
 
-void Music::setDuration(unsigned long durationTemp) {
+void Music::setDuration(int durationTemp) {
   duration = durationTemp;
 }
 
@@ -20,7 +20,7 @@ char* Music::getArtist() {
   return artist;
 }
 
-unsigned long Music::getDuration() {
+int Music::getDuration() {
   return duration;
 }
